+++ CHANGELOG.md
@@ -8,6 +8,9 @@ Unreleased section should follow [Release Toolkit](https://github.com/newrelic/r
 
 ## Unreleased
 
+### Enhancement
+- Support setting up CLI values through environment variables
+
 ## v0.5.4 - 2026-09-21
 
 ### 🐞 Bug fixes